Refactor file layout
This moves the transport code to a new file so it will no longer run on
init for now.
