+++ th/constants.py
@@ -1,0 +1,7 @@
+HELP = """
+Use the following keys:
+
+- A, S, D, F, and G to press fret buttons.
+- Space bar to pick string.
+- X to quit
+"""
