+++ th/screens.py
@@ -1,0 +1,5 @@
+from .controller import GameController
+
+
+def stage(screen, game_state):
+    screen.play([GameController(screen, game_state)], stop_on_resize=True)
