+++ th/state.py
@@ -1,0 +1,7 @@
+class GameState(object):
+    """
+    Persistent state for this application.
+    """
+
+    def __init__(self):
+        self.button_1 = False
